tests: add server tests
