+++ packages/code-du-travail-frontend/package.json
@@ -96,7 +96,8 @@
     "stylelint-config-recommended": "^3.0.0",
     "stylelint-config-styled-components": "^0.1.1",
     "stylelint-order": "^4.1.0",
-    "stylelint-processor-styled-components": "^1.10.0"
+    "stylelint-processor-styled-components": "^1.10.0",
+    "supertest": "^4.0.2"
   },
   "jest": {
     "testMatch": [
